01.11.2020

+++ RequestsService/DTO/User/SignUpInputDTO.cs
@@ -58,6 +58,6 @@ namespace RequestsService.DTO.User
         /// <summary>
         /// Факультет
         /// </summary>
-        public int FacultyID { get; set; }
+        public long FacultyID { get; set; }
     }
 }
